Update 3 KL MyAI: go up, down, forward, turning...
AI now can go up and down (might need some debugging), go forward, turn left and right, all while updating curDir and curRow and curCol. I keep track of the movements but not the coordinates so that still needs to be worked on. The backtracking took a lot of time to do but at least that part works for all movements I believe. The movement backtracking may become obsolete after we finish coding the directional movements and store all the coordinates.