Update 4
MyAI can now move in all four cardinal directions without any difficulties. (Up, down, left, right). However I do not check for bounds (there is no check for bump, some other code must be written so that goUp is not called or the game board should be updated so that goUp can return an exception or something if it is at the top already.
Finished debugging turning left and right issues
To demonstrate this, MyAI goes up 2 coordinates, 
    goes back down 2, 
    goes right 1, 
    left 1, 
    climb out